pyproject.toml
[project]
name = "tools-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Yodan", email = "u9223259@bsmch.net" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.0",
    "fastmcp>=3.0.2",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langfuse>=3.14.5",
    "langgraph>=1.0.10",
    "numpy>=2.2.6",
    "openai>=2.24.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["uv_build>=0.10.7,<0.11.0"]
build-backend = "uv_build"
